功能：查询题目详细信息

+++ jc-club-subject/jc-club-subject-infra/src/main/java/com/jingdianjichi/subject/infra/basic/entity/SubjectMultiple.java
@@ -26,7 +26,7 @@ public class SubjectMultiple implements Serializable {
     /**
      * 选项类型
      */
-    private Long optionType;
+    private Integer optionType;
     /**
      * 选项内容
      */
